+++ source/kernel/mem/pmm.c
@@ -172,8 +172,8 @@ void pmm_initRegion(uint32_t base, size_t size) {
     serialPrintf("pmm_initRegion: Region at 0x%x (size 0x%x) initialized. Blocks used: %i. Free blocks: %i\n", base, size, blocks, (pmm_maxBlocks - pmm_usedBlocks));
 }
 
-// pmm_deinitRegion(uint32_t base, size_t size) - Marks a region as unusable memory (e.g. kernel region)
-void pmm_deinitRegion(uint32_t base, size_t size) {
+// pmm_deinitRegion(uintptr_t base, size_t size) - Marks a region as unusable memory (e.g. kernel region)
+void pmm_deinitRegion(uintptr_t base, size_t size) {
     // maybe bugged? dunno
     int align = base / 4096;
     int blocks = size / 4096;
