+++ source/sysroot/usr/include/kernel/vesa.h
@@ -30,16 +30,16 @@ typedef struct {
 	uint16_t segment_a;
 	uint16_t segment_b;
 	uint32_t win_func_ptr;		// deprecated; used to switch banks from protected mode without returning to real mode
-	uint16_t pitch;			// number of bytes per horizontal line
-	uint16_t width;			// width in pixels
+	uint16_t pitch;				// number of bytes per horizontal line
+	uint16_t width;				// width in pixels
 	uint16_t height;			// height in pixels
-	uint8_t w_char;			// unused...
-	uint8_t y_char;			// ...
+	uint8_t w_char;				// unused...
+	uint8_t y_char;				// ...
 	uint8_t planes;
-	uint8_t bpp;			// bits per pixel in this mode
-	uint8_t banks;			// deprecated; total number of banks in this mode
+	uint8_t bpp;				// bits per pixel in this mode
+	uint8_t banks;				// deprecated; total number of banks in this mode
 	uint8_t memory_model;
-	uint8_t bank_size;		// deprecated; size of a bank, almost always 64 KB but may be 16 KB...
+	uint8_t bank_size;			// deprecated; size of a bank, almost always 64 KB but may be 16 KB...
 	uint8_t image_pages;
 	uint8_t reserved0;
  
@@ -61,7 +61,7 @@ typedef struct {
 
 
 // Framebuffers (probably not a good idea but I could care less)
-extern uint32_t *vbeBuffer;
+extern uint8_t *vbeBuffer;
 extern uint32_t *framebuffer; // 2nd framebuffer.
 extern uint32_t modeWidth;
 extern uint32_t modeHeight;
